Add case for when recv DMA eth packet has 0 data length
If the function receive_frame_C receives a DMA Ethernet packet that has a 0
useful data length it now returns the length of the padding and pointer to the
start of the payload.

I do not know why we need this but without it the code does not work. Nor do I
know how we can get a DMA Ethernet packet without data.