Fix send packet length and padding
When sending Ethernet packet with Ethertype the length of payload was being
over written by data.

Moved the "runt" packet check (frame length smaller then min size) to the send
packet function as padding should not be included as part of the payload data
length.

Removed the "payload size divisible by 4" checkout when sending data as I don't
see the need for it when sending to the DMA port.