Write the tbsp_send_reset function
This reset function will attempt to reset the hardware 10 times, testing to see
if it was successful (by checking that the seq pos is zero). It needs to be
done multiple time in case the reset packet is dropped or if the socket buffers
have not been flushed from last usage (can happen in Linux) and we recv old
packets. The number 10 was chosen as I have observed 8 stale packets in the
socket buffers under Linux.