Write the tbsp_send_stream function
Sending is "optimistic" in that the entire data block is sent and then we check
to see if it was sent successfully. This is most efficient assuming few dropped
packets but if many packets get dropped we could see the same data reset many
times. Under that condition it would be more efficient to send fewer bytes
before the handshake.