Add sleep in tbsp reset function
This is after the reset has been sent and before the "req" packet gets sent. It
gives the pico board time to come out of reset and handle the req packet. This
speeds up the reset as the before the req packet was being dropped and we were
waiting on the timeout ack to be sent.