rm reset from loadBytecode function in NeuFlow
And add explicit sendReset() to all the demos. Having the reset in the load
function did not give enough time to the openflower to come out of booting and
for it to reset the streamer port before the byte code began to be sent. This
caused the tbsp stream to fail early and thus the bytecode would have to be
resent many times.