New tbsp header format, have 2 seq # in data packet
The data and request packet types now has both the tx and rx sequence numbers
in the header and the ack packets revise the order of its seq numbers (so that
they are the same as the data packet). This allows for the stream to be synced
even if ack packets are lost or a data packet is received instead.