Use user reg alloc in Core & Interface
The functions in Core no longer uses fixed user register addresses for operations.

The etherflow Interface claims reg_A and reg_B as it uses both for some
operations. Those registers are not freed as I have not had time to determine
what each operation does and what would be the best way to dynamically allocate
the registers for each function.