Add a loopRepeat to replace the other loopRepeats
Instead of having to call two functions to do a repeat loop we now only have to
do one. This new repeat function takes as an argument the number of repeats and
a function containing the code that will be repeated over. This new function
also handles if zero or negative numbers are passed in as arguments. If it is a
negative number it loops infinitely and if zero it does not include the code at
all.