Add loopUntilZero to Core, a one function until loop
It has at least two Arguments the first, *mode* is true or false depending on
if it loops until reg is zero or until it is not zero. The second, *code* is
the loop body to be executed/looped over in the form of a function.  The code
MUST return the register that is going to be tested. Any other arguments are
passes to the code function when it is exacted.