Bug fix Camera running buf port addr's & sending
Passing in a ref to neuflow object when initializing the Camera object instead
of just the core so all other objects are available.

Correct addressing of the streamer ports when being reconfigured to do
continues writes for the camera.

Have the running buffer get image function mirror the copyToHost functions
functionality.