Use BPF (Barkeley Packet Filter) instead of raw socket on OSX
On OSX, the BPF interface is now used. It allows to launch the software without being root.
The delay in the network_send_packet() function has been increase from 10us to 100us for OSX to improve connection stability. It seems that it need to be properly tuned for each machine.