BPF in etherflow, dynamic sleep delay and custom network interface
Etherflow is modified to use the Berkeley packet filter for OSX. It allows to run software without being root.

A dynamic delay is now used to wait between two sent frames. If the computer is too slow and the processing time greater than the needed delay between the two frames, it doesn't wait.
The usleep function is calibrated because it shows an important bias.

We can now specify a custom interface name to use with the NeuFlow. This name is specified at the creation of the NeuFlow object.