Remove all bootloader.content references
The 'bootloader.content' was pre-compiled bootloader code that was written for
a very differently configured NeuFlow system. It had been superseded by the
Core:bootSequence(args) function but was still being added to the start of
every demo bytecode we ran.

Might have been causing problems when the NeuFlow system was changed in ways
incompatible with the 'bootloader.content'.