Add 2 process version of face-detector
This version of the face-detector is almost exactly the same as the other
face-detector demo except it uses the torch package 'parallel' to fork the
camera init into a child process. This child process reads an image from the
camera and sends it to the parent process to be processed and displayed.