Enable non kernel packing for allocEmbeddedData
By default incoming data is reformatted & treated as a kernel. If non kernel
data needs to be embedded an explicit 1D or 2D packing argument needs to be
passed in. If 2D is selected but the width of the data is larger then the
streamer (memory) stride, packing is reverted to 1D.