Replace sys register allocator
The new register allocator uses weak tables to track in use registers. In use
registers are kept in the weak table only so long as an outside variable keeps
it accessible. Once it is no longer accessible the lua garbage collection
removes it from the weak table. This removes the need to manually use a "free"
method in the allocator.