Remove the binary table from Core
The table to store binary data (to be included into the bytecode) has been
removed from Core and is now passed in as an argument to the functions that
process and transform the data. This was done so the binary table no longer
needed to be cleared by the Core:startProcess function.