Replace send packet delay with static value in tbsp
Sending an Ethernet packet requires a small delay between sending another
packet. There was a bit of code in the tbsp library that would measure the time
between sending packets and only usleep the minimum amount to ensure the
minimum delay is reached. This code was causing a much to long delay and so has
(temporarily?) been replaced with a static value which has been tested to work.