Add <on-board-processing> profiler ack for tbsp
First pass at adding a profiling step for the on board processing. It estimates
the time needed to process the convnet by measuring the time it takes from
having sent an input and being ready to receive the output.

The assumption being made is that the demo will transfer data to the device in
the following manner:-

nf:copyToDev(inputs)
nf:copyFromDev(outputs)

If data transfer does not look like this, the profiling step will not work.