Add old pico camera control code
This camera control code was developed against a very old version of
master. As such it does not allocate openflower registers or memory
correctly. This prevents it from working.

It is imported as a first step to updating it.

For a proper history of the commits leading up to its development you
should look at the 'cameras' branch.